Improve code and unit tests
